requirements.txt
acres==0.5.0
aiocache==0.12.3
aiocqhttp==1.4.4
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.14.0
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.5.3
attrs==25.4.0
audioop-lts==0.2.2
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.8
cloudscraper==1.2.71
colorama==0.4.6
colorful==0.5.6
configobj==5.0.9
configparser==7.2.0
contourpy==1.3.1
cssselect==1.2.0
cyaron==0.6.0
cycler==0.12.1
dearpygui==2.0.0
discord.py==2.6.4
docopt==0.6.2
etelemetry==0.3.1
feedparser==6.0.12
filelock==3.19.1
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Moment==1.0.6
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.55.3
frozenlist==1.8.0
git-filter-repo==2.47.0
h11==0.14.0
h2==4.1.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
Hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
kiwisolver==1.4.7
libgravatar==1.0.4
looseversion==1.3.0
luogu==0.1.0
lxml==5.3.0
Mako==1.3.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.4
MouseInfo==0.1.3
multidict==6.7.0
natsort==8.4.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
nonebot==1.9.1
numpy==2.2.0
ollama==0.4.7
packaging==24.2
pandas==2.3.2
parsel==1.9.1
pefile==2024.8.26
pillow==11.0.0
pipreqs==0.4.13
pipx==1.7.1
platformdirs==4.3.6
priority==2.0.0
propcache==0.4.1
prov==2.1.1
puremagic==1.30
PyAutoGUI==0.9.54
pydantic==2.10.6
pydantic_core==2.27.2
pydot==4.0.1
PyGetWindow==0.0.9
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
PyMsgBox==2.0.1
PyMuPDF==1.26.3
pynput==1.8.1
pyparsing==3.2.0
pyperclip==1.10.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytweening==1.2.0
pytz==2025.2
pywin32-ctypes==0.2.3
pyxnat==1.6.3
PyYAML==6.0.2
Quart==0.20.0
rarfile==4.2
rdflib==7.1.4
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
scipy==1.16.1
setuptools==80.9.0
sgmllib3k==1.0.0
simple-websocket==1.1.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
tqdm==4.67.1
traits==7.0.2
tuack==0.1.5.3.3
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
userpath==1.9.2
w3lib==2.2.1
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
wxPython==4.2.3
xeger==0.4.0
yarg==0.1.10
yarl==1.22.0
